(issue 46) fix %{key} not being substituted
This commt implements a fix for issue #46

Before this commit, using %{key} in the uri(s) didn't get interpolated
with the key name being looked up. This used to be the behaviour with
hiera v3.

The reason for this is that the new uri(s) options now get interpolated
by expand_uris in location_resolver, even before it reaches the
backend. That meants that by the time the call reaches the hiera
backend, %{key} is no longer there.

I worked around this by implementing a simple string replacement that
looks for __KEY__ and replaces it with the key name.

To verify this change, use `__KEY__` as part of the uri or uris options.
The request should include the name of the key instead of `__KEY__`