Added json_key option
Currently, when using "output": "json", the function will always to lookup the full key in the `json` returned by the HTTP endpoint. However, in most cases you do not want that and setting `output` to something else will require you to parse the JSON data in Puppet.

This small change will allow you to set json_key to false if you want to return the full JSON object to Hiera, or set it to a custom key that is, in turn, interpolated with the "special tags".

For example, using this function as a Hiera backend for Consul:

hierarchy:
  - name: "Consul backend"
    uris:
      - "http://127.0.0.1:8500/v1/__KEY__?raw"
    lookup_key: hiera_http
    options:
      ignore_404: true
      output: json
      json_key: false

$ echo -n '{"ip":"127.0.0.1","name":"Another test!"}' | consul kv put test/data1 -
$ puppet lookup test/data1
---
ip: 127.0.0.1
name: Another test!
$ puppet lookup test/data1.ip
--- 127.0.0.1
...
$ puppet lookup catalog/services
---
consul: []
...
$ puppet lookup catalog/nodes.0.Datacenter
--- dc1
...